Split module parsing and types out better
