+++ smol-core/src/Smol/Core/Modules/Parser.hs
@@ -1,15 +1,24 @@
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE OverloadedStrings #-}
 
-module Smol.Core.Parser.Module
-  ( moduleParser,
+
+module Smol.Core.Modules.Parser
+  (
+  moduleParser,
+
+
+    parseModule,
+    parseModuleAndFormatError,
   )
 where
+
 
 import qualified Data.List.NonEmpty as NE
 import Data.Text (Text)
 import Data.Void
 import Smol.Core.Modules.Types.ModuleItem
+import Smol.Core.Modules.Types.TestName
 import Smol.Core.Parser.DataType (dataTypeParser)
 import Smol.Core.Parser.Expr
 import Smol.Core.Parser.Identifiers
@@ -19,8 +28,27 @@ import Smol.Core.Parser.Typeclass
 import Smol.Core.Typecheck.Typeclass.Types
 import Smol.Core.Types
 import Text.Megaparsec hiding (parseTest)
+import Smol.Core.Parser.Primitives (textPrim)
+import Data.Bifunctor (first)
+import qualified Data.Text as T
+import Text.Megaparsec.Char
 
 type Parser = Parsec Void Text
+
+type ParseErrorType = ParseErrorBundle Text Void
+
+parseAndFormat :: Parser a -> Text -> Either Text a
+parseAndFormat p = first (T.pack . errorBundlePretty) . parse (p <* eof) "repl"
+
+parseModule :: Text -> Either ParseErrorType [ModuleItem Annotation]
+parseModule = parse (space *> moduleParser <* eof) "repl"
+
+parseModuleAndFormatError :: Text -> Either Text [ModuleItem Annotation]
+parseModuleAndFormatError = parseAndFormat (space *> moduleParser <* eof)
+
+-------
+
+
 
 -- currently fails at the first hurdle
 -- since we can parse each thing separately, maybe
@@ -181,3 +209,6 @@ parseClass = do
             tcFuncType = ty
           }
       )
+
+testNameParser :: Parser TestName
+testNameParser = myLexeme $ TestName <$> textPrim
