+++ smol-core/src/Smol/Core/Modules/Types.hs
@@ -1,7 +1,6 @@
 module Smol.Core.Modules.Types
   ( module Smol.Core.Modules.Types.Entity,
     module Smol.Core.Modules.Types.Module,
-    module Smol.Core.Modules.Types.ModuleName,
     module Smol.Core.Modules.Types.ModuleItem,
     module Smol.Core.Modules.Types.TestName,
     module Smol.Core.Modules.Types.DefIdentifier,
@@ -14,7 +13,6 @@ import Smol.Core.Modules.Types.DefIdentifier
 import Smol.Core.Modules.Types.Entity
 import Smol.Core.Modules.Types.Module
 import Smol.Core.Modules.Types.ModuleItem
-import Smol.Core.Modules.Types.ModuleName
 import Smol.Core.Modules.Types.Test
 import Smol.Core.Modules.Types.TestName
 import Smol.Core.Modules.Types.TopLevelExpression
