+++ smol-core/src/Smol/Core/Modules/Types/Entity.hs
@@ -9,7 +9,7 @@ module Smol.Core.Modules.Types.Entity where
 -- terrible, pls improve
 import Data.Aeson (FromJSON, FromJSONKey, ToJSON, ToJSONKey)
 import GHC.Generics (Generic)
-import Smol.Core.Modules.Types.ModuleName
+import Smol.Core.Types.ModuleName
 import Smol.Core.Printer
 import Smol.Core.Types.Constructor
 import Smol.Core.Types.Identifier
