+++ smol-core/src/Smol/Core/Parser.hs
@@ -6,9 +6,7 @@ module Smol.Core.Parser
     parseExprAndFormatError,
     parseTypeAndFormatError,
     parseType,
-    parseModule,
     parseConstraint,
-    parseModuleAndFormatError,
     parseDataTypeAndFormatError,
     parseConstraintAndFormatError,
     ParseErrorType,
@@ -19,10 +17,8 @@ import Data.Bifunctor (first)
 import Data.Text (Text)
 import qualified Data.Text as T
 import Data.Void
-import Smol.Core.Modules.Types
 import Smol.Core.Parser.DataType (dataTypeParser)
 import Smol.Core.Parser.Expr
-import Smol.Core.Parser.Module
 import Smol.Core.Parser.Type
 import Smol.Core.Parser.Typeclass
 import Smol.Core.Typecheck.Types
@@ -46,14 +42,8 @@ parseExpr = parse (space *> expressionParser <* eof) "repl"
 parseExprAndFormatError :: Text -> Either Text ParserExpr
 parseExprAndFormatError = parseAndFormat (space *> expressionParser <* eof)
 
-parseModule :: Text -> Either ParseErrorType [ModuleItem Annotation]
-parseModule = parse (space *> moduleParser <* eof) "repl"
-
 parseConstraint :: Text -> Either ParseErrorType (Constraint ParseDep Annotation)
 parseConstraint = parse (space *> constraintParser <* eof) "repl"
-
-parseModuleAndFormatError :: Text -> Either Text [ModuleItem Annotation]
-parseModuleAndFormatError = parseAndFormat (space *> moduleParser <* eof)
 
 parseDataTypeAndFormatError :: Text -> Either Text (DataType ParseDep Annotation)
 parseDataTypeAndFormatError = parseAndFormat (space *> dataTypeParser <* eof)
