+++ smol-core/test/Main.hs
@@ -1,7 +1,6 @@
 module Main (main) where
 
 import Test.Hspec
-import qualified Test.Interpreter.InterpreterSpec
 import qualified Test.Modules.CheckSpec
 import qualified Test.Modules.FromPartsSpec
 import qualified Test.Modules.InterpreterSpec
@@ -9,6 +8,7 @@ import qualified Test.Modules.PrettyPrintSpec
 import qualified Test.Modules.ResolveDepsSpec
 import qualified Test.Modules.RunTestsSpec
 import qualified Test.Modules.TypecheckSpec
+import qualified Test.Modules.ParserSpec
 import qualified Test.ParserSpec
 import qualified Test.TransformSpec
 import qualified Test.Typecheck.ExhaustivenessSpec
@@ -29,7 +29,6 @@ main = hspec $ parallel $ do
   Test.Typecheck.TypeclassSpec.spec
   Test.Typecheck.ToDictionaryPassingSpec.spec
   Test.ParserSpec.spec
-  Test.Interpreter.InterpreterSpec.spec
   Test.Modules.CheckSpec.spec
   Test.Modules.FromPartsSpec.spec
   Test.Modules.InterpreterSpec.spec
@@ -37,4 +36,5 @@ main = hspec $ parallel $ do
   Test.Modules.ResolveDepsSpec.spec
   Test.Modules.RunTestsSpec.spec
   Test.Modules.TypecheckSpec.spec
+  Test.Modules.ParserSpec.spec
   Test.TransformSpec.spec
