+++ smol-core/test/Test/Modules/PrettyPrintSpec.hs
@@ -6,6 +6,7 @@
 -- | test evaluating and running tests for a module
 module Test.Modules.PrettyPrintSpec (spec) where
 
+import qualified Smol.Core.Modules.Parser as Parse
 import Data.Bifunctor (second)
 import Data.FileEmbed
 import Data.Foldable (traverse_)
@@ -14,7 +15,6 @@ import Data.Text (Text)
 import qualified Data.Text.Encoding as T
 import Smol.Core.Modules.PrettyPrint (printModuleParts)
 import Smol.Core.Modules.Types.ModuleItem
-import qualified Smol.Core.Parser as Parse
 import Smol.Core.Printer
 import Test.Hspec
